requirements.txt
streamlit>=1.24.0
pillow>=9.0.0
openai>=1.0.0